MainActivity is now a Navigation Drawer
title
also i fixed the LGHSTwitterFeedActivity so that it works now for ppl
without the twitter app
