+++ app/src/main/res/layout/content_main.xml
@@ -162,6 +162,4 @@
         </LinearLayout>
     </ScrollView>
 
-
-
 </android.support.constraint.ConstraintLayout>
